requirements.txt
aiolimiter==1.1.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
openai==1.16.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
tzlocal==5.2
